pyproject.toml
[tool.poetry]
name = "growbot-server"
version = "2023.0"
description = "Bucket-based, code-powered, 3d-printed horticulture"
authors = ["Damien Allen <dev@dallen.co>"]

[tool.poetry.dependencies]
python = "^3.11"

asyncio-mqtt = "^0.16.1"
boto3 = "^1.26.108"
botocore = "^1.29.108"
fastapi = "^0.92.0"
pandas = "^1.5.3"
python-decouple = "^3.8"
tqdm = "^4.65.0"
uvicorn = "^0.20.0"

influxdb-client = {extras = ["async", "ciso"], version = "^1.36"}

[tool.poetry.dev-dependencies]
black = "^23.1"
ipython = "^8.10"
ipykernel = "^6.21"
mypy = "^1.0"
notebook = "^6.5"
ruff = "^0.0.252"

[build-system]
requires = ["poetry>=1.3"]
build-backend = "poetry.masonry.api"
